TS-984 Move the code from utils into the newSchoolSubSection to fix the build error for codegen
